+++ engines/query-sparql/test/util-browser.js
@@ -1,0 +1,5 @@
+const fetchFn = globalThis.fetch;
+
+export function fetch(...args) {
+    return fetchFn(...args);
+}
